Now compiles on windows. Added dummy versions of assembly functions, must be replaced to be thread safe!!!!

+++ src/pieces/shared_data.h
@@ -55,7 +55,7 @@ private:
     SharedData(const SharedData&);
     SharedData& operator=(const SharedData&);
 
-    unsigned int m_count;
+    unsigned long m_count;
 };
 
 } // namespace Pieces
